Version [a.0.0] - First Alpha Release!
FloorRenderer now sets only lit AND seen tiles to player view.

Cursor mode now describes what's at the cursor in the target window.

Added arrival and completion lore for Dark Grove.

Added history and myth lore.

Implemented non-crashing player death and victory.

Added new material to support wolf and bear armor.

+++ va2/src/engine/Engine.java
@@ -68,6 +68,8 @@ public class Engine extends Saveable {
                 actors.remove(a);
             deadActors = new ArrayList<>();
             for (Iterator<Actor> i = actors.iterator(); i.hasNext();) {
+                if (deadActors.contains(Session.getPlayer().getActor()))
+                    return; //immediately terminate engine handling if the player has died.
                 actor = i.next();
                 if (gameTurn % STATUS_INTERVAL == 0) //apply all relevant status checks before the death check, in case they kill the actor.
                     handleStatus(actor);
